// frontend/app/src/app/core/api/model/vehicleDataQualityResponse.ts
/**
 * FS Bus API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Counts and match flags explaining which sources contributed to the page.
 */
export interface VehicleDataQualityResponse { 
    matched_vehicle_master: boolean;
    matched_smart_fleet: boolean;
    matched_trip_data: boolean;
    matched_bi_data: boolean;
    matched_events: boolean;
    matched_inspections: boolean;
    trip_count: number;
    event_count: number;
    inspection_count: number;
    bi_score_count: number;
}