// frontend/app/src/app/core/api/model/vehicleDetailResponse.ts
/**
 * FS Bus API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { VehicleCurrentStatusResponse } from './vehicleCurrentStatusResponse';
import { VehicleDataQualityResponse } from './vehicleDataQualityResponse';
import { VehicleInspectionHistoryResponse } from './vehicleInspectionHistoryResponse';
import { VehicleResponse } from './vehicleResponse';
import { VehicleTripDetailResponse } from './vehicleTripDetailResponse';
import { VehicleEventDetailResponse } from './vehicleEventDetailResponse';
import { VehicleScorePointResponse } from './vehicleScorePointResponse';


/**
 * Full vehicle drilldown payload.
 */
export interface VehicleDetailResponse { 
    vehicle: VehicleResponse;
    current_status: VehicleCurrentStatusResponse;
    trips: Array<VehicleTripDetailResponse>;
    events: Array<VehicleEventDetailResponse>;
    score_points: Array<VehicleScorePointResponse>;
    inspection_history: Array<VehicleInspectionHistoryResponse>;
    data_quality: VehicleDataQualityResponse;
}