// frontend/app/src/app/core/api/model/vehicleScorePointResponse.ts
/**
 * FS Bus API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * One plotted score point for a vehicle trip graph.
 */
export interface VehicleScorePointResponse { 
    label: string;
    trip_start?: string | null;
    style_score?: number | null;
    route_score?: number | null;
}